requirements.txt
fastapi
uvicorn
chromadb
pinecone
pgvector
psycopg2-binary
langgraph
langchain
langchain-google-genai
langchain-groq
crewai
pandas
